Correct debian dependencies
Add in pkg-config, since that's used at build time to figure out where
libs are located.

We depend on libuv, so that needs to be installed!

libmd is for our numpy module, which is partially linked towards it at
build time. Should probably reconsider this, perhaps include, just like
we do for libuv etc or add as run time dep?

For installing from tar ball we no longer depend on libprotobuf-c-dev &
libutf8proc-dev since those are part of the acton distribution.