Use container for building
Having problems with rust in the ubuntu-latest VM in GitHub Actions
that I can't reproduce elsewhere on Ubuntu 20.04 (what ubuntu-latest
currently maps to) so making the environment more deterministic by using
a container.

We're on bookworm because 11 (bullseye) doesn't have a new enough rust /
cargo for like anything? to work.