Update hero headline and USPs
I think this - "Distributed cloud systems" - more accurately describes
the intended use case for Acton.  We are trying to build fault tolerant
distributed systems, primarily because we want the fault tolerance. We
will also get the distributed model from that so that it should be
possible to write scalable systems that scale out horizontally, however,
the primary use case isn't really for "massive scale". Like if someone
wants to build planet-scale systems like Google's bla bla then they are
probably better of using languages and databases explicitly which favors
explicit control, at the expense of development time. Acton should be
simple. Writing a Google service will never be simple. Different goals,
different solutions.

Also, I think we are targeting mission critical, so wanted to throw
that in there. It's a good phrase. Also "programming platform", which
feels like a compiler and more.. which is what it is.

wip